+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ./src/index.js",
-    "dev": "nodemon ./src/index.js"
+    "start": "node ./src/server.js",
+    "dev": "nodemon ./src/server.js"
   },
   "repository": {
     "type": "git",
@@ -23,12 +23,15 @@
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
+    "ejs": "^3.1.6",
+    "eslint-config-trybe-backend": "^1.0.4",
     "express": "^4.17.1",
     "express-rescue": "^1.1.31",
     "jsonwebtoken": "^8.5.1",
     "pg": "^8.6.0"
   },
   "devDependencies": {
+    "eslint": "^7.31.0",
     "nodemon": "^2.0.7"
   }
 }
