+++ src/app.js
@@ -1,0 +1,15 @@
+const cors = require('cors');
+const express = require('express');
+const bodyParser = require('body-parser');
+
+const routes = require('./routes');
+
+const app = express();
+
+app.use(cors());
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+
+app.use(routes);
+
+module.exports = app;
