+++ src/routes/admin/index.js
@@ -1,0 +1,3 @@
+const routes = require('express').Router();
+
+module.exports = routes;
