+++ src/routes/api/loginRoute.js
@@ -1,7 +1,7 @@
 const routes = require('express').Router();
 const rescue = require('express-rescue');
 
-const login = require('../controllers/loginController');
+const login = require('../../controllers/loginController');
 
 routes.post('/', rescue(login.create));
