+++ src/routes/api/usersRoute.js
@@ -1,7 +1,7 @@
 const routes = require('express').Router();
 const rescue = require('express-rescue');
 
-const users = require('../controllers/usersController');
+const users = require('../../controllers/usersController');
 
 routes.post('/', rescue(users.create));
 routes.get('/:id', rescue(users.findById));
