+++ src/server.js
@@ -1,0 +1,6 @@
+require('dotenv').config();
+const app = require('./app');
+
+const { PORT } = process.env;
+
+app.listen(PORT, () => console.log(`Listening on port ${PORT}`));
